fix: hook userCache into settings.tsx to help propagate changes to components who need them
