+++ src/context/settings.tsx
@@ -19,7 +19,7 @@ import useActivityTracker from '@/hooks/useActivityTracker'
 import { INTERVALS } from '@/utils/time'
 import axios from 'axios'
 import * as https from 'node:https'
-import useUserCache from '@/hooks/useUserCache'
+import { USER_CONFIG_CACHE } from '@/types/app_params'
 
 const RETRY_ATTEMPTS = 3
 
@@ -177,18 +177,69 @@ interface ISettingsConfig {
   setPriorityFee?: Dispatch<SetStateAction<PriorityFeeName>>
   latency: number
   priorityFeeValue: number
+  userCache: USER_CONFIG_CACHE
+  setUserCache: (cache: USER_CONFIG_CACHE) => void
+  updateUserCache: (cache: Partial<USER_CONFIG_CACHE>) => void
 }
 
 const SettingsContext = React.createContext<ISettingsConfig | null>(null)
+
+function newCache(): USER_CONFIG_CACHE {
+  return {
+    hasDexOnboarded: false,
+    farm: {
+      hasFarmOnboarded: false,
+      showDepositedFilter: false
+    },
+    gamma: {
+      hasGAMMAOnboarded: false,
+      showDepositedFilter: false
+    },
+    hasSignedTC: false,
+    endpointName: 'QuickNode',
+    endpoint: null,
+    priorityFee: 'Default'
+  } as USER_CONFIG_CACHE
+}
+
+export function resetUserCache(): void {
+  window.localStorage.setItem('gfx-user-cache', JSON.stringify(newCache()))
+}
+
+export function getOrCreateCache(): USER_CONFIG_CACHE {
+  const rawCache = window.localStorage.getItem('gfx-user-cache')
+  if (rawCache) {
+    try {
+      return JSON.parse(rawCache) as USER_CONFIG_CACHE
+    } catch (e) {
+      console.error('Error parsing user cache', e)
+    }
+  }
+  const cache = newCache()
+  localStorage.setItem('gfx-user-cache', JSON.stringify(cache))
+  return cache
+}
+
+export function validateUserCache(cache?: USER_CONFIG_CACHE): boolean {
+  const validCache = cache ?? getOrCreateCache()
+  const validCacheKeys = Object.keys(validCache)
+  const emptyValidCache = newCache()
+  for (const key of validCacheKeys) {
+    if (!(key in emptyValidCache)) {
+      return false
+    }
+    if (typeof validCache[key] !== typeof emptyValidCache[key]) {
+      return false
+    }
+  }
+  return true
+}
 
 export function useSlippageConfig(): {
   slippage: number
   setSlippage: React.Dispatch<React.SetStateAction<number>>
 } {
   const context = useContext(SettingsContext)
-  if (!context) {
-    throw new Error('Missing settings context')
-  }
 
   const { slippage, setSlippage } = context
   return { slippage, setSlippage }
@@ -196,9 +247,6 @@ export function useSlippageConfig(): {
 
 export function useConnectionConfig(): ISettingsConfig {
   const context = useContext(SettingsContext)
-  if (!context) {
-    throw new Error('Missing settings context')
-  }
 
   return context
 }
@@ -209,12 +257,22 @@ export const SettingsProvider: FC<{ children: ReactNode }> = ({ children }) => {
   const [slippage, setSlippage] = useState<number>(DEFAULT_SLIPPAGE)
   const [blacklisted, setBlacklisted] = useState<boolean>(false)
   const [isUnderMaintenance, setIsUnderMaintenance] = useState<boolean>(false)
-  const {userCache, updateUserCache} = useUserCache();
+  const [userCache, setUserCache] = useState<USER_CONFIG_CACHE>(getOrCreateCache())
   const [endpointName, setEndpointName] = useState<EndPointName>(userCache.endpointName || 'QuickNode')
   const [priorityFee, setPriorityFee] = useState<PriorityFeeName>(userCache.priorityFee || 'Default')
   const [latency, setLatency] = useState<number>(0)
   const [shouldTrack, setShouldTrack] = useState<boolean>(true)
+  const setCache = useCallback((cache: USER_CONFIG_CACHE) => {
+    setUserCache(cache)
+  }, [])
 
+  const updateUserCache = useCallback((cache: Partial<USER_CONFIG_CACHE>) => {
+    console.log('TOS USER CACHE UPDATE', cache)
+    setUserCache((prevCache) => ({ ...prevCache, ...cache }))
+  }, [])
+  useEffect(() => {
+    window.localStorage.setItem('gfx-user-cache', JSON.stringify(userCache))
+  }, [userCache])
   const { priorityFeeInstruction, priorityFeeValue } = useMemo(() => {
     let fee = 0.0
     switch (priorityFee) {
@@ -278,7 +336,8 @@ export const SettingsProvider: FC<{ children: ReactNode }> = ({ children }) => {
     updateUserCache({
       endpointName: endpointName,
       endpoint: endpointName === CUSTOM_RPC.name ? endpoint : null,
-      priorityFee});
+      priorityFee
+    })
 
     // creates connection - temp ws url
     return new Connection(endpoint, {
@@ -358,7 +417,10 @@ export const SettingsProvider: FC<{ children: ReactNode }> = ({ children }) => {
         setPriorityFee,
         priorityFeeInstruction,
         latency,
-        priorityFeeValue
+        priorityFeeValue,
+        userCache,
+        setUserCache: setCache,
+        updateUserCache
       }}
     >
       {children}
