+++ src/pages/FarmV3/ChoosePool.tsx
@@ -5,10 +5,9 @@ import 'styled-components/macro'
 import Slider from 'react-slick'
 import { checkMobile } from '../../utils'
 import { Pool, poolType } from './constants'
-import { useDarkMode, useSSLContext } from '../../context'
+import { useConnectionConfig, useDarkMode, useSSLContext } from '../../context'
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
-import useUserCache from '@/hooks/useUserCache'
 
 const STYLED_POPUP = styled(PopupCustom)<{
   currentSlide: number
@@ -176,7 +175,7 @@ export const ChoosePool: FC<{
   const [isError, setIsError] = useState<boolean>(false)
   const sliderRef = useRef<any>()
   const { mode } = useDarkMode()
-  const {userCache, updateUserCache} = useUserCache();
+  const {userCache, updateUserCache} = useConnectionConfig();
 
   const [userAnswer, setUserAnswer] = useState<any>({
     answerOne: null,
